@@ -10,9 +10,9 @@
   ],
   "scripts": {
     "prestart": "npm install",
-    "start": "cd df-client-app && npm start",
-    "build": "cd df-client-app && npm run build",
-    "build:development": "cd df-client-app && npm run build:development",
+    "start": "npm run --workspace=df-client-app start",
+    "build": "npm run --workspace=df-client-app run build",
+    "build:development": "npm run --workspace=df-client-app run build:development",
     "clean": "rm -rf node_modules; rm -rf **/node_modules; rm -rf **/dist; rm **/tsconfig.tsbuildinfo",
     "lint": "npm run lint --workspaces --if-present",
     "lint:css": "npm run lint:css --workspaces --if-present",
@@ -35,6 +35,7 @@
     "typescript-eslint": "^8.12.1"
   },
   "dependencies": {
+    "@trussworks/react-uswds": "^10.0.2",
     "@typescript-eslint/eslint-plugin": "^5.59.6",
     "@typescript-eslint/parser": "^5.59.6",
     "dotenv-expand": "^10.0.0",
@@ -55,6 +56,6 @@
     "nanoid": "^3.3.8"
   },
   "engines": {
-    "node": "18.20.4"
+    "node": "22.18.0"
   }
 }
