@@ -1,25 +1,25 @@
-import homer from './mocks/homer.json';
-import marge from './mocks/marge.json';
-import lisa from './mocks/lisa.json';
-import lisaKnockout from './mocks/lisa_knockout.json';
-import bart from './mocks/bart.json';
-import sean from './mocks/sean.json';
-import abe from './mocks/abe.json';
-import grandma from './mocks/grandma.json';
-import milhouse from './mocks/milhouse.json';
-import sideshowbob from './mocks/sideshowbob.json';
-import krusty from './mocks/krusty.json';
-import margeWithOnlySadiSuccess from './mocks/marge_with_only_sadi_success.json';
-import ned from './mocks/ned.json';
-import december27 from './mocks/december27.json';
-import kent from './mocks/kent.json';
-import jon from './mocks/jon.json';
-import margeWithSpaces from './mocks/marge__simpson.json';
-import w2ParseError from './mocks/w2_parse_error.json';
-import carly from './mocks/carly.json';
-import w2_with_decimal_values from './mocks/w2_with_decimal_values.json';
-import maude from './mocks/maude.json';
-import rod from './mocks/rod.json';
+import homer from '../../../../../../backend/src/main/resources/dataimportservice/mocks/homer.json';
+import marge from '../../../../../../backend/src/main/resources/dataimportservice/mocks/marge.json';
+import lisa from '../../../../../../backend/src/main/resources/dataimportservice/mocks/lisa.json';
+import lisaKnockout from '../../../../../../backend/src/main/resources/dataimportservice/mocks/lisa_knockout.json';
+import bart from '../../../../../../backend/src/main/resources/dataimportservice/mocks/bart.json';
+import sean from '../../../../../../backend/src/main/resources/dataimportservice/mocks/sean.json';
+import abe from '../../../../../../backend/src/main/resources/dataimportservice/mocks/abe.json';
+import grandma from '../../../../../../backend/src/main/resources/dataimportservice/mocks/grandma.json';
+import milhouse from '../../../../../../backend/src/main/resources/dataimportservice/mocks/milhouse.json';
+import sideshowbob from '../../../../../../backend/src/main/resources/dataimportservice/mocks/sideshowbob.json';
+import krusty from '../../../../../../backend/src/main/resources/dataimportservice/mocks/krusty.json';
+import margeWithOnlySadiSuccess from '../../../../../../backend/src/main/resources/dataimportservice/mocks/marge_with_only_sadi_success.json';
+import ned from '../../../../../../backend/src/main/resources/dataimportservice/mocks/ned.json';
+import december27 from '../../../../../../backend/src/main/resources/dataimportservice/mocks/december27.json';
+import kent from '../../../../../../backend/src/main/resources/dataimportservice/mocks/kent.json';
+import jon from '../../../../../../backend/src/main/resources/dataimportservice/mocks/jon.json';
+import margeWithSpaces from '../../../../../../backend/src/main/resources/dataimportservice/mocks/marge__simpson.json';
+import w2ParseError from '../../../../../../backend/src/main/resources/dataimportservice/mocks/w2_parse_error.json';
+import carly from '../../../../../../backend/src/main/resources/dataimportservice/mocks/carly.json';
+import w2_with_decimal_values from '../../../../../../backend/src/main/resources/dataimportservice/mocks/w2_with_decimal_values.json';
+import maude from '../../../../../../backend/src/main/resources/dataimportservice/mocks/maude.json';
+import rod from '../../../../../../backend/src/main/resources/dataimportservice/mocks/rod.json';
 
 import { DataImportProfile } from './dataImportProfileTypes.js';
 import { DataImportRootResponseSchema } from './schema/DataImportServiceResponse.js';
