@@ -1,13 +1,14 @@
 import { FactGraph } from '@irs/js-factgraph-scala';
 import { Condition, RawCondition } from '../flow/Condition.js';
-type HasConditions = { condition?: RawCondition; conditions?: RawCondition[] };
+import { ConfigProps, FlowComponentName, FlowNodeType } from '../flow/ContentDeclarations.js';
+import { ComponentProps } from 'react';
 
-export const conditionsPass = (props: HasConditions, fg: FactGraph, collectionId: string | null) => {
+export const conditionsPass = (props: ConfigProps<ComponentProps<FlowNodeType<FlowComponentName>>>, fg: FactGraph, collectionId: string | null) => {
   if (props && props.condition) {
     return new Condition(props.condition).evaluate(fg, collectionId);
     // eslint-disable-next-line eqeqeq
   } else if (props.conditions != undefined) {
-    return props.conditions.every((c) => {
+    return props.conditions.every((c: RawCondition) => {
       const result = new Condition(c).evaluate(fg, collectionId);
       return result;
     });
