@@ -1,4 +1,4 @@
-import { UserConfig, defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config';
 import react from '@vitejs/plugin-react';
 import browserslistToEsbuild from 'browserslist-to-esbuild';
 import { JSON_SCHEMA } from 'js-yaml';
@@ -7,7 +7,7 @@ import viteTsconfigPaths from 'vite-tsconfig-paths';
 import autoprefixer from 'autoprefixer';
 import path from 'path';
 
-export const configOptions: UserConfig = {
+export default defineConfig({
   base: process.env.DF_CLIENT_PUBLIC_PATH || `/df/file`,
   assetsInclude: [`**/*.svg`],
   plugins: [react(), viteTsconfigPaths(), ViteYaml({ schema: JSON_SCHEMA })],
@@ -17,7 +17,7 @@ export const configOptions: UserConfig = {
       localsConvention: `camelCaseOnly`,
     },
     postcss: {
-      plugins: [autoprefixer()],
+      plugins: [autoprefixer() as any],
     },
     preprocessorOptions: {
       scss: {
@@ -97,7 +97,4 @@ export const configOptions: UserConfig = {
       },
     },
   },
-};
-
-// https://vitejs.dev/config/
-export default defineConfig(configOptions);
+});
